requirements.txt
pandas
matplotlib
seaborn
python-docx
plotly